+++ public/components/common/modules/panel/components/agg-table.tsx
@@ -1,5 +1,18 @@
+/*
+ * Wazuh app - React component Aggregations Table.
+ *
+ * Copyright (C) 2015-2021 Wazuh, Inc.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * Find more information about this on the LICENSE file.
+ */
+
 import { EuiBasicTable, EuiPanel, EuiTitle, EuiBasicTableColumn } from '@elastic/eui';
-import { useEsSearch } from '../hooks';
+import { useEsSearch } from '../../../hooks';
 import React from 'react';
 
 export const AggTable = ({
@@ -9,7 +22,7 @@ export const AggTable = ({
   maxRows,
   tableTitle,
   panelProps,
-  titleProps
+  titleProps,
 }) => {
   const preAppliedAggs = {
     buckets: {
@@ -20,9 +33,9 @@ export const AggTable = ({
       },
     },
   };
-  const  {esResults, isLoading, error}  = useEsSearch({ preAppliedAggs });
+  const { esResults, isLoading, error } = useEsSearch({ preAppliedAggs });
   const buckets = ((esResults.aggregations || {}).buckets || {}).buckets || [];
-  const columns:EuiBasicTableColumn<any>[] = [
+  const columns: EuiBasicTableColumn<any>[] = [
     {
       field: 'key',
       name: aggLabel,
@@ -49,7 +62,13 @@ export const AggTable = ({
       <EuiTitle {...titleProps}>
         <h2>{tableTitle}</h2>
       </EuiTitle>
-      <EuiBasicTable items={buckets} columns={columns} rowProps={getRowProps} loading={isLoading} error={error ? error.message : undefined} />
+      <EuiBasicTable
+        items={buckets}
+        columns={columns}
+        rowProps={getRowProps}
+        loading={isLoading}
+        error={error ? error.message : undefined}
+      />
     </EuiPanel>
   );
 };