+++ public/components/common/modules/panel/index.ts
@@ -1,6 +1,5 @@
 /*
- * Wazuh app - Office 365 Panel index.
- *
+ * Wazuh app - Panel components
  * Copyright (C) 2015-2021 Wazuh, Inc.
  *
  * This program is free software; you can redistribute it and/or modify
@@ -11,4 +10,5 @@
  * Find more information about this on the LICENSE file.
  */
 
-export * from './office-panel';
+export { MainPanel } from './main-panel';
+export * from './components/';